Add V4L2 MJPEG encoder and example
* Split out H264 encoder from V4L2 encoder
* Also fix broken imports on video examples
